Backport ffmpeg 8.0 build fixes to 25.08 branch
Building with ffmpeg 8.0 was fixed in d39fae995420d5b78d40a51a5a748c1ecd74a0ee by @amazingakai however that commit also raised the minimum ffmpeg required to build ffmpegthumbs. As we can't increase version requirements in release branches that commit cannot be cherry-picked as-is.

This commit instead takes the relevant changes from that commit and wraps them in version check ifdefs. That allows us to fix the build with ffmpeg 8.0 but without breaking compatibility with older ffmpeg version. This code is ugly but it shouldn't matter since it will never hit any branch besides the 25.08 one.

LIBAVCODEC_VERSION_MAJOR 60 corresponds to ffmpeg 6.x for the record.